Move `StorePickUp`, `Payment` and `Shipping` component folders
