+++ react/components/OrderInfo/OrderInfo.js
@@ -3,11 +3,11 @@ import PropTypes from 'prop-types'
 import parcelify from '@vtex/delivery-packages'
 import OrderHeader from './OrderHeader'
 import CustumerInfo from './CustumerInfo'
-import PaymentSummary from './Payment/PaymentSummary'
-import Shipping from './Shipping/Shipping'
+import PaymentSummary from '../Payment/PaymentSummary'
+import Shipping from '../Shipping/Shipping'
 import OrderTotal from './OrderTotal'
 import OrderSummary from './OrderSummary'
-import StorePickUp from './StorePickUp/StorePickUp'
+import StorePickUp from '../StorePickUp/StorePickUp'
 
 import { profileShape } from '../../proptypes/shapes'
